+++ src/const.js
@@ -61,21 +61,5 @@ const RankToTextContent = {
   MOVIE_BUFF: 'Movie Buff',
 };
 
-const StatisticPeriodValue = {
-  ALL_TIME: 'all-time',
-  TODAY: 'today',
-  WEEK: 'week',
-  MONTH: 'month',
-  YEAR: 'year',
-};
-
-const StatisticPeriodLabel = {
-  ALL_TIME: 'All time',
-  TODAY: 'Today',
-  WEEK: 'Week',
-  MONTH: 'Month',
-  YEAR: 'Year',
-};
-
-export {StatisticPeriodLabel, StatisticPeriodValue, RankToTextContent, EMOTIONS, NEW_COMMENT_DEFAULT, SortType, KeyboardKey, UserAction, UpdateType, FilterType, EventType, Screen, Rank};
+export {RankToTextContent, EMOTIONS, NEW_COMMENT_DEFAULT, SortType, KeyboardKey, UserAction, UpdateType, FilterType, EventType, Screen, Rank};
 export default KeyboardKey;
